Fix NFS mocking in snapd and apparmor tests.
